v1.0.1
spring security

+++ src/main/java/com/bigzkoop/cboapi/dto/DebitDetails.java
@@ -4,23 +4,18 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class DebitDetails {
     private Long debitDetId;
     private Long debitId;
-    private Long totalAmount;
-    private Long basicAmount;
-    private Long totalInstallments;
-    private Long paidInstallments;
-    private Long amountPerInstallments;
-    private Long remainingInstallments;
-    private Long remainingAmount;
+    private Double totalAmount;
+    private Double basicAmount;
+    private Double totalInstallments;
+    private Double paidInstallments;
+    private Double amountPerInstallments;
+    private Double remainingInstallments;
+    private Double remainingAmount;
 
 }
